+++ mrbuilder/__init__.py
@@ -1,0 +1,21 @@
+try:
+    import torch
+    from .builders.pytorch import \
+        get_model_builder, \
+        get, \
+        get_model, \
+        load, \
+        build, \
+        register_layer_builders
+except ImportError as e:
+    try:
+        import tensorflow as tf
+        from .builders.keras import \
+            get_model_builder, \
+            get, \
+            get_model, \
+            load, \
+            build, \
+            register_layer_builders
+    except ImportError as e:
+        raise RuntimeError("No Valid Backends Found") from e