+++ tests/base/test_bootstrap.py
@@ -8,39 +8,31 @@ class TestBootstrapBase:
 
     @classmethod
     def get_num_layers(cls, model) -> int:
-        raise NotImplemented
+        raise NotImplementedError
 
     @classmethod
     def get_layer(cls, model, i):
-        raise NotImplemented
+        raise NotImplementedError
 
     @classmethod
     def get_type(cls, model, i):
-        raise NotImplemented
+        raise NotImplementedError
 
     @classmethod
     def get_stride(cls, model, layer_num, dimension=0):
-        raise NotImplemented
+        raise NotImplementedError
 
     @classmethod
     def get_size(cls, model, i):
-        raise NotImplemented
+        raise NotImplementedError
 
     @classmethod
-    def get_in_size(cls, model, i):
-        raise NotImplemented
+    def get_in_channels(cls, model, i):
+        raise NotImplementedError
 
     @classmethod
-    def get_out_size(cls, model, i):
-        raise NotImplemented
-
-    @classmethod
-    def get_in_size(cls, model, i):
-        raise NotImplemented
-
-    @classmethod
-    def get_out_size(cls, model, i):
-        raise NotImplemented
+    def get_out_channels(cls, model, i):
+        raise NotImplementedError
 
 
 class TestBuilderBase: